renamed default locale file to en-US.json
