+++ crowdin.yml
@@ -1,3 +1,3 @@
 files:
-  - source: /src/locales/default.json
+  - source: /src/locales/en-US.json
     translation: /src/locales/%locale%.json
